[Core] CheckColdStakeFreeOutput: skip SPORK checks if mnsync incomplete
Currently CheckBlock skips the masternode (or budget) payment check in coinstakes when IsInitialBlockDownload() returns true (essentially when the current best block time is older than nMaxTipAge seconds).

Since, in order to verify the block payee, the masternode list needs to be up to date, IsBlockPayeeValidalso skips the checks when mnsync has not finished.
CheckColdStakeFreeOutput however, does not, so, when the spork list is not updated, SPORK_13 and SPORK_9 cannot be verified (in case of budget payment blocks) resulting in CheckBlock failures with

ERROR: CheckColdStakeFreeOutput: Wrong cold staking outputs:

(Unless the user doesn't manually set a lower nMaxTipAge to force IsInitialBlockDownload to true).

This fixes it.