Don't throw when generate doesn't create all blocks (PoS)
Just return the list of hashes for the blocks generated.
This affects only "live" usage of regtest wallet, as the functional tests use the utility function generate_pos which calls generate(1) and increases mocktime (or adds a sleep) when the node can actually stake but can't find a valid kernel at the current time.