nStakeSplitThreshold: division by zero
RPC should accept only values between 1 and 999999 (as per setstakesplitthreshold help) but the input provided was only checked against the upper bound.
Setting a 0 value would result in a crash (division-by-zero) at the first stake.

Now prevent the wallet from splitting the stake if the threshold is set to zero.