CBlockIndex cleanup
Removes a number of (unused or barely used) class variables/methods in CBlockIndex, and introduces a way to use a single variable to store either v1 or v2 stake modifiers.

Since CDiskBlockIndex serialization starts with CLIENT_VERSION, this pr  dont bumps the CLIENT_VERSION_BUILD in order to introduce a new serialization without requiring reindex.
Version is set so 9.0.99. if we want use it change version