Sync Upgrades, part 2
Effort to bring us up-to-date with upstream's sync.h/cpp sources.

The only task that left to be fully up-to-date with upstream, that will leave for a third PR because it touches other areas of the sources and want to keep this as small as possible, is the CCriticalSection removal and replacement with the RecursiveMutex typedef.