Remove reserve balance
reservebalance is a leftover from peercoin code which was probably never used.
Currently it is also broken: the final loop in CWallet::StakeableCoins is useless (updating an un-used local variable nAmountSelected and not removing any utxo from the vector of stakeable coins).

Failing to see a valid use-case for it (not already provided by the stake-split threshold), I propose to just remove it.

This also makes the call to GetStakingBalance from StakeableCoins not needed (thus one less loop over mapWallet txes).