SendMultiRow: clear address label when address is changed
In the send tab, when there is only one recipient, the address label is automatically loaded by the wallet if the user enters a known address in the recipient field.

Problem: If the user overwrites the recipient entry, inserting a new address, and this address is not found in the addressbook (or even simply deletes the previous inserted address), then the label does not change, and keeps displaying the one fetched earlier (the label of the previous address).
This of course could lead to issues of users sending to a different recipient than intended.

Fix: Trivial. clear the label whenever the address is changed (before updating it if the address inserted is labeled in the addressbook).