Remove IsSuperMajority
IsSuperMajority checks are slow, and not needed after enforcement.
Also ISM is not a good mechanism for deployment, so will not likely be used again after BIP65.
(based on bitcoin#8391)

We should implement better fork signal/activation solutions such as BIP 9 in the future.

Note: testnet resync completed. currently checking resync on mainnet.