Fix masternodecurrent: return next winner, not rank-1 mn
masternodecurrent RPC is improperly returning the masternode with rank 1 instead of the next winner.

Note: GetCurrentMasterNode returns the masternode with lower rank and it's used as winner only when no valid block payee can be found. We should probably change its name to something more meaningful (GetLowestRankMasterNode ?) and change the comment calling it "winner". Same goes for the comment above CMasternode::CalculateScore (masternodes are not even paid based on their scores, but only on the time of their last payment).