Upgrade to HD wallet
The rationality behind this work is that locked wallets, as them are encrypted, are not able to upgrade to HD wallet. The keystore and thereby keys are not available to perform any action at initialization time.

This PR solves it adding a new JSON-RPC command upgradewallet to be able to upgrade the wallet at runtime, requesting the wallet passphrase if it's needed.

A functional test verifying the new command was included inside the wallet_upgrade.py test.