Run tests on pre-HD wallets
This adds a startup flag -legacywallet, which, on first run, enables the creation of a pre-HD wallet (when a wallet.dat is already present, the flag is just ignored). This flag is available for RegTest only.

A special option --legacywallet is also added to the functional test_runner.
Single tests (or the entire test suite) can now be performed over legacy wallets easily, e.g.:

./test_runner.py -j10 --legacywallet

./wallet_basic.py --legacywallet

Note 1: Running with pre-HD wallets, the wallet_accounts.py test fails. It is fixed in a successive PR.

Note 2: Travis currently is set to run the functional test suite only on HD wallets.