Restore macOS Travis tests
This PR restores the macOS TravisCI tests by way of a (somewhat clunky) workaround that forces a homebrew update. The TravisCI build images are outdated, providing a no longer supported version of homebrew. Hopefully this workaround is only needed temporarily.

Reference of the issue can be found Here.

At the same time, I've added ccache support to CMake based builds, if ccache is available. Local build times are greatly benefited by this change, but TravisCI builds seem to be largely unaffected (could be a caching issue) .

Note: The changes to the OpenSSL paths for macOS are primarily due to the fact that homebrew no longer provides OpenSSL 1.0.x packages. These changes are likely NOT backwards compatible with an environment that has frozen/retained the (now defunct) 1.0.x homebrew packages.