Add Unlock/Relock context flow
Fix UnlockContext class and connect it where needed.
Now, once an action requires the wallet to be fully unlocked, the askpassphrase dialog is presented to the user (if the status is "Locked" or "UnlockedForStaking"). This way, the password can be entered directly, and the operation can resume without needing additional clicks by the user.
The previous locking status is restored when the action completes.

Interested flows:

    Send transaction
    Create New Receive Address
    Create New Payment Request
    Create Masternode Controller
    Create Cold Staking Address
    Send Cold Stake Delegation
    Wallet Data: Add multisend recipient
    Tools: Sign Message
    Tools: BIP38 tool
    Upgrade wallet to HD