Fix min depth requirement for stake inputs in AvailableCoins
If nDepth is equal to nStakeMinDepth - 1 then the coin is not skipped in AvailableCoins and gets included in the list of stakeable utxos.
This utxo won't be able to stake though, as it will fail ContextCheck in Stake() (logging an error), thus it shouldn't have been included in the first place.