CheckColdStakeFreeOutput when mnsync not complete
Problem: if the finalized budgets and the masternode list are not synchronized, the output returned from IsBudgetPaymentBlock is unreliable.

For the moment we fix it by checking that the sum of the stake input value and block reward is at least equal to the sum of the values of non-free outputs.

A better approach, for the future, would be probably to outright prevent any call to CheckBlock when IsInitialBlockDownload is false but mnSync is not fully completed yet.