Fix g_best_block initialization in ABC
g_best_block holds a cached best block hash.
This is updated during ActivateBestChain (in UpdateTip).

Problem: at startup, the object is not initialized until a new block arrives (ActivateBestChainStep is not called) thus g_best_block != chainActive.Tip()->GetBlockHash() so, if the client stakes before receiving a new block, that stake gets orphaned.
In case of segregated devnet with just one staker, it is impossible to move the chain at all.