Do not print an error on connection timeouts through proxy
If a timeout happens while reading the proxy response, this effectively means we timed out while connecting to the remote node. This is very common for Tor, so do not print an error message to the log in that case.

Avoids a flood of

ERROR: Error reading proxy response

while running through Tor, causing users to blame the proxy.