Stake Modifier V3
New stake modifier protocol (v3) proposed as hard fork.
Instead of using a chain of hashes, we propose to switch to a chain of signatures.
So the new modifier, instead of being the hash of the previous modifier, combined with the txId of the coinstake input, becomes the hash of a signature, performed with the key of the coinstake prevout, over the previous modifier.

This makes the modifier completely unpredictable, thus offering a much better protection against grinding.

The flow is fairly simple: the staker, after finding a valid kernel, signs the modifier used, and publishes the signature (prefixed by a new opcode OP_STAKEMODIFIER) in the first output of the coinstake transaction. All peers then verify the signature during the proof of stake checks. The hash of this signature is saved in the block index, and becomes the modifier to be used for the successive block's kernel.

Note: the new modifier is enforced bumping the block version (to V8).

Mainnet uses a placeholder for now.