Fix wrong Lock/Unlock zWallet method call
Fixing a wrongly implemented lock/unlock process for the zWallet object and correcting the wallet parent setup in zWallet and CzPIVTracker objects which was bad too.

The previous flow was calling pwalletMain from crypter when the wallet is actually a subclass of it.. Blocking any unit test possibility over the wallet class.

1)pwalletMain can or cannot be created.
2) It can get to the point of more than one Wallet object created (example, unit tests creating simple wallets with different db paths that have nothing to do with pwalletMain) and the CzPIVWallet/CzPIVTracker pointing to only one of them.