ClientModel cacheTip sync with back-end
Fixing cacheTip not sync with the back-end information when the node has no connection and thereby there was no tip signal broadcasted to the front-end to update the cached tip. Showing a wrong chain height instead of the correct back-end tip.

This will only happen one time. Once the tip gets updated, it will not request it anymore from the back-end and only receive signals. Not locking cs_main anymore.