Customize the timeout of the SnackBar based on its message length
The default 3-seconds timeout for the snackbar is not optimal in all the situations.
Sometimes the message is longer, and the user has not enough time to read it before it hides.
Other times the message is very short (e.g. "Wallet Unlocked"), and 3 seconds seem a bit much.

This PR proposes to use a variable timeout, based on the length of the message to display (50 milliseconds per character), with a minimum set at 2 seconds and a maximum at 10.