Spend cold-stake delegations
Problem:
A few users have reported of having accidentally spent cold-stake delegations, after interacting with the GUI, in the following cases

    Sending a transaction
    Creating a MN controller with the masternode wizard
    Creating another delegation

Cause:
Cold stakes and delegations are picked up (owner-side) by default in AvailableCoins (in order to get the correct balances).
While in the RPC functions it is possible to use a boolean parameter fIncludeDelegations to decide whether or not to let the automatic selection algorithm include them (if needed), the GUI has no such option, thus, by default, doesn't skip them.

The user is presented with a warning before spending a cold-stake delegation, but some user doesn't pay attention, and, in any case, dismissing and retrying would result again in the same warning.

The only way for the user to be sure not to include any delegation is using coin control.
But this is tricky as well, since P2CS outputs are not clearly distinguished there, the user can only identify them by checking the owner address (this particular aspect will be addressed in a future PR).

Further, coin control is not available in the MN Wizard, so this doesn't solve 2.

Proposed solution:

    A new checkbox in the send widget lets the user decide whether to include delegated coins in the selection algorithm. It is un-checked by default.