[GUI] MasternodeWizard validations
fixes the masternode alias: right now only letters and digits are allowed, but special characters (except spaces and hash marks) should be allowed too. It also replaces deprecated QRegExp with QRegularExpression and fixes the port number to the default.

 introduces validation for masternode IP addresses: they should be valid IPv4, IPv6 or .onion addresses and, if not on regtest, they should be reachable and routable. A unit test for the new validateMasternodeIP function is added to netbase_tests.
Also removes the validation for the port, since it's fixed to network-default.