[GUI][Bug] Reset custom change address
First commit fixes a bug where the btnChangeAddress in SendWidget was deactivated, despite a custom address still being saved (e.g. saving a valid address, and then trying to change it to an invalid one).

Second commit addresses a feature request: being able to reset only the change address (without resetting the whole coin control) directly from within the dialog.
With this PR, when a custom change address is set, the text of the "CANCEL" button in SendChangeAddressDialog is changed to "RESET" and, when clicked, resets the default change address to CNoDestination (which means that a new change address will be created during the send operation).
The upright "X" button retains its original function of just closing the dialog.