[Bug][Wallet] Fix insane fees
nFeeNeeded is an absolute value (feePerKb * txKb), and, as explained both in GUI labels, and in the RPC help, -payTxFee is a value per Kb.
CWallet::GetMinimumFee , instead, was improperly setting nFeeNeeded to just feePerKb when a user had a custom fee set, resulting in the creation of transactions with fee much higher than expected (most transactions are well under 1Kb in size).