[Bug] Simpler nTimeSmart computation
Given the strict timestamp checks enforced with POS TimeProtocol v2 (introduced with  # 1002: a block cannot have a timestamp earlier than previous blocks and cannot be more than 15 seconds in the future), we don't need the complex (and error-prone) logic of the current implementation of ComputeTimeSmart.
We can, instead, simply rely on the blocktime, which already ensures a proper ordering for the transactions.
This also allows us to encapsulate the function in the CWalletTx class (renaming it to UpdateTimeSmart).

We can also get rid of the additional GetComputedTxTime (with its additional LOCK, and main chain lookup), and just use GetTxTime.

Finally, update the nSmartTime when a mempool transaction is added to a block.