[Core] Prevector
Based on:

[Core] Big endian support

This introduces prevector<N, T>, a new basic data type which is a fully API compatible drop-in replacement for std::vector<T> and uses it for the script, to reduce the considerable memory overhead of vectors in cases where they normally contain a small number of small elements.

Original tests in Bitcoin showed a reduction in dbcache memory usage by 23%, and made an initial sync 13% faster.

Backported from the following upstream's PRs, with only additional edits in the 2nd layer scripts (masternode-payments, masternode-budget) and in the unit tests (to address the differences in insecure_rand):

    bitcoin # 6914
    bitcoin # 7888
    bitcoin # 8850
    bitcoin # 9349
    bitcoin # 9505
    bitcoin # 9856
    bitcoin # 10534
    bitcoin # 11011
    bitcoin # 14028