Prepare for non-Base58 addresses [Step 1]
This patch removes the need for the intermediary Base58 type CBitcoinAddress, by providing {Encode,Decode,IsValid} Destination functions that directly operate on the conversion between std::strings and CTxDestination.

As a side, it also fixes a number of indentation issues, and removes probably several unnecessary implicit CTxDestination<->CBitcoinAddress conversions.

It's not a small change but it will give us much more flexibility once it's fully done over every current and new address type.

The first commit is coming from upstream partially ( 11117 ), the rest is purely tackling our code (we have so so many more places using the CBitcoinAddress class).

And finally, have implemented a second abstraction (Destination wrapper) on top of the base58 address to {Encode,Decode,IsValid} Staking addresses as well.

Note:
This migration is not fully complete. It's first step, we still have places using the CBitcoinAddress object instead of the CTxDestination abstraction.