[Refactoring] Lightweight abstraction of boost::filesystem
Based on top of:

[DB] Bitcoin 0.12-0.14 dbwrapper improvements

Backports bitcoin # 9902

Wrap boost::filesystem as the fs namespace, contained in one header.
Also add fsbridge for bridging between stdio.h fopen() and the path type.

This allows:

Replacing it when the time comes with std::filesystem (when standardized, and when we start using the appropriate c++ version) with a single change.

Using custom filesystem implementations that offer the same interface. I need this in cloudabi, a sandboxed environment which has no concept of a global filesystem namespace, where paths need to be accompanied by a file descriptor which grants access to it.

It's less typing, too.