[Core] Implement accurate memory accounting for mempool
Based on top of:

Implement accurate memory accounting for mempool

This continues the work of #1531 adding memory accounting for the mempool.
Backports bitcoin # 6410.

Original description:

This implements accurate memory usage accounting for the mempool. It is only exposed through getmempoolinfo for now, but could be used for limiting the resource requirements too (bitcoin # 6281).