[BUG][GUI] check for available "unlocked" balance in prepareTransaction
If the user tries to send an amount exceeding his balance, the wallet correctly returns the error related to AmountExceedsBalance or AmountWithFeeExceedsBalance.
Although, if the wallet has enough coins, but some of them are locked, the transaction preparation will fail, without giving a specific reason.
This happens because prepareTransaction checks the amount to send against the total wallet's balance (without excluding locked coins if any).
We fix this introducing a getUnlockedBalance() function in the model, and using it in prepareTransaction.

Similarly, in the Send Widget, the "total remaining" shows the wallet's balance (locked + unlocked), therefore the transaction might fail if the user tries to send it all, while having locked coins.
We will use the new getUnlockedBalnce() here too and fix the label (when there is no coin selected via coin control).