[Wallet][RPC] FundTransaction - fundrawtransaction
based on top of

[Cleanup] Remove obsolete multisig code

This introduces a new wallet function, CWallet::FundTransaction() (and exposes it via RPC with fundrawtransaction), to fill a tx containing only vouts (or not enough vins to cover the vouts) with unspent coins from the wallet.

fundrawtransaction will not modify existing inputs, and will add one change output (if needed) to the outputs. It will not sign the inputs (so can include also watch-only or multi-sig inputs, if enabled).

backported from:

    bitcoin # 6088
    bitcoin # 17219 [*]
    bitcoin # 6417
    bitcoin # 6444
    bitcoin # 6415
    bitcoin # 6828
    bitcoin # 7296 (only bebe58b)
    bitcoin # 7506
    bitcoin # 7518
    bitcoin # 7967

adapting the tests for the (more recent) framework.

[*] Note: this has been included to be able to call fundrawtransaction without the need for an unencrypted wallet (for the change address key)