Move zerocoin validation to its own legacy file
This is a move-only PR, no functionality or code was changed.

The main idea behind this decoupling is continue cleaning main.cpp. Making new functionalities and back ports easier to implement.

Have been thinking whether this should be a new file or zerocoin_verify.h and ended with a new file because this flows are one step above the verify methods and are not isolated from the process that calls them. Example: AcceptToMemoryPoolZerocoin works only for the AcceptToMemoryPool method.